+++ react-ui/src/common/useVoting.jsx
@@ -4,53 +4,97 @@
  * votingUnavailable: votingDisabled || votingExpired
  */
 
-import differenceInMilliseconds from 'date-fns/differenceInMilliseconds';
-import isFuture from 'date-fns/isFuture';
-import { useEffect, useState } from 'react';
+import { deleteData, putData } from '../api';
+import useContestId from '../utils/useContestId';
 
-import useSwrContest from '../utils/useSwrContest';
-
+import snackbarTypes from './snackbarTypes';
+import useAuthState from './useAuthState';
 import useComponentsState from './useComponentsState';
+import useSnackbarState from './useSnackbarState';
+import useSwrMutation from './useSwrMutation';
+import useVotingStatus from './useVotingStatus';
 
-const MAX_DELAY = 2147483647;
+const VOTES_URL = '/votes';
 
-function useVoting() {
-  const [{ votingDisabled }, setComponentsState] = useComponentsState();
-  const [votingExpired, setVotingExpired] = useState(false);
+const updateEntries = (entries, { entryId, rating }) => entries.reduce((acc, cur) => {
+  const id = cur.imgurId ?? cur.id;
+  if (id !== entryId) {
+    acc.push(cur);
+  } else {
+    acc.push({ ...cur, rating });
+  }
+  return acc;
+}, []);
 
-  const {
-    data: {
-      voteEnd,
+function useVoting(entryId) {
+  const contestId = useContestId();
+  const contestUrl = `/contests/${contestId}`;
+  const setComponentsState = useComponentsState()[1];
+  const [{ isLoggedIn }] = useAuthState();
+  const { disableVoting } = useVotingStatus();
+
+  const { trigger: triggerPut } = useSwrMutation(
+    contestUrl,
+    putData,
+    VOTES_URL,
+  );
+  const { trigger: triggerDelete } = useSwrMutation(
+    contestUrl,
+    deleteData,
+    VOTES_URL,
+  );
+
+  const updateSnackbarState = useSnackbarState();
+  const showError = () => {
+    updateSnackbarState(snackbarTypes.VOTING_ERROR);
+  };
+  const triggerOptions = (input) => ({
+    optimisticData: (current) => ({ ...current, entries: updateEntries(current.entries, input) }),
+    revalidate: false,
+    populateCache: (response, contest) => {
+      if (!response) {
+        showError();
+        return contest;
+      }
+
+      const newEntries = updateEntries(contest.entries, input);
+      const newData = { ...contest, entries: newEntries };
+      updateSnackbarState(snackbarTypes.VOTING_SUCCESS);
+      return newData;
     },
-    mutate,
-  } = useSwrContest();
+    onError: () => {
+      showError();
+    },
+  });
 
-  const voteEndDate = new Date(voteEnd);
-
-  useEffect(() => {
-    if (isFuture(voteEndDate)) {
-      const timeToEnd = Math.min(differenceInMilliseconds(voteEndDate, new Date()), MAX_DELAY)
-      const timeout = setTimeout(() => {
-        mutate();
-        setVotingExpired(true);
-      }, timeToEnd);
-      return () => clearTimeout(timeout);
+  const changeRating = (newValue, oldValue) => {
+    if (!isLoggedIn) {
+      setComponentsState({ redditLogInDialogOpen: true });
+      return;
     }
 
-    setVotingExpired(true);
-    return () => { };
-  }, [voteEndDate]);
+    if (newValue === oldValue) {
+      return;
+    }
 
-  const votingUnavailable = votingDisabled || votingExpired;
-  const disableVoting = (isDisabled) => {
-    setComponentsState({ votingDisabled: isDisabled });
+    disableVoting(true);
+
+    const voteInput = { contestId, entryId, rating: newValue };
+    triggerPut(voteInput, triggerOptions(voteInput));
+  };
+
+  const clearRating = () => {
+    disableVoting(true);
+
+    const input = { contestId, entryId };
+    triggerDelete(input, triggerOptions(input));
   };
 
   return {
-    voteEndDate,
-    votingDisabled,
-    votingUnavailable,
-    disableVoting,
+    changeRating,
+    clearRating,
+    triggerDelete,
+    triggerPut,
   };
 }
