+++ react-ui/src/components/VotingSlider.jsx
@@ -7,18 +7,14 @@ import Slider from '@material-ui/core/Slider';
 import { makeStyles, withStyles } from '@material-ui/core/styles';
 import clsx from 'clsx';
 import PropTypes from 'prop-types';
-import { useState, useEffect, useRef } from 'react';
-import { useParams } from 'react-router-dom';
+import { useEffect, useRef } from 'react';
 
-import { deleteData, putData } from '../api';
 import {
-  useAuthState, useComponentsState, useSnackbarState, useSwrMutation, useVoting,
+  useVoting, useVotingStatus,
 } from '../common';
-import snackbarTypes from '../common/snackbarTypes';
 
 const MIN_SCORE = 0;
 const MAX_SCORE = 5;
-const VOTES_URL = '/votes';
 
 const ThemedSlider = withStyles((theme) => ({
   root: {
@@ -84,40 +80,13 @@ const useStyles = makeStyles((theme) => ({
   },
 }));
 
-const updateEntries = (entries, { entryId, rating }) => entries.reduce((acc, cur) => {
-  const id = cur.imgurId ?? cur.id;
-  if (id !== entryId) {
-    acc.push(cur);
-  } else {
-    acc.push({ ...cur, rating });
-  }
-  return acc;
-}, []);
-
 function VotingSlider({
   entryId, rating,
 }) {
-  const setComponentsState = useComponentsState()[1];
-  const { disableVoting, votingUnavailable: disabled } = useVoting();
-  const [{ isLoggedIn }] = useAuthState();
-  const { contestId } = useParams();
+  const { votingUnavailable: disabled } = useVotingStatus();
   const ratingRef = useRef(rating);
-  const [isInteractive, setInteractive] = useState(false);
 
-  const contestUrl = `/contests/${contestId}`;
-  // useSwrContest???
-  const { isMutating: isMutatingPut, trigger: triggerPut } = useSwrMutation(
-    contestUrl,
-    putData,
-    VOTES_URL,
-  );
-  const { isMutating: isMutatingDelete, trigger: triggerDelete } = useSwrMutation(
-    contestUrl,
-    deleteData,
-    VOTES_URL,
-  );
-
-  const updateSnackbarState = useSnackbarState();
+  const { changeRating, clearRating } = useVoting(entryId);
 
   const classes = useStyles();
 
@@ -125,56 +94,8 @@ function VotingSlider({
     ratingRef.current = rating;
   }, [rating]);
 
-  useEffect(() => {
-    if (isInteractive) {
-      disableVoting(isMutatingDelete || isMutatingPut);
-    }
-  }, [isInteractive, isMutatingDelete, isMutatingPut]);
-
-  const showError = () => {
-    updateSnackbarState(snackbarTypes.VOTING_ERROR);
-  };
-
-  const triggerOptions = (input) => ({
-    optimisticData: (current) => ({ ...current, entries: updateEntries(current.entries, input) }),
-    revalidate: false,
-    populateCache: (response, contest) => {
-      if (!response) {
-        showError();
-        return contest;
-      }
-
-      const newEntries = updateEntries(contest.entries, input);
-      const newData = { ...contest, entries: newEntries };
-      updateSnackbarState(snackbarTypes.VOTING_SUCCESS);
-      return newData;
-    },
-    onError: () => {
-      showError();
-    },
-  });
-
   const handleSliderChange = async (event, newValue) => {
-    if (!isLoggedIn) {
-      setComponentsState({ redditLogInDialogOpen: true });
-      return;
-    }
-
-    if (newValue === ratingRef.current) {
-      return;
-    }
-
-    setInteractive(true);
-
-    const voteInput = { contestId, entryId, rating: newValue };
-    triggerPut(voteInput, triggerOptions(voteInput));
-  };
-
-  const clearRating = () => {
-    setInteractive(true);
-
-    const input = { contestId, entryId };
-    triggerDelete(input, triggerOptions(input));
+    changeRating(newValue, ratingRef.current);
   };
 
   const isUnrated = !rating && rating !== 0;
